[SERVICE-543] Moved EventSpecification definition to client, to avoid client importing from provider. Prevented EventRouter from appearing in API docs. Can use number keys in demo app to create/close notifications. Moved ActionTrigger definition to actions file. Increased some test delays to account for toast animations.
