+++ src/client/internal.ts
@@ -9,9 +9,9 @@
  * This file is excluded from the public-facing TypeScript documentation.
  */
 
-import {NotificationActionResult} from './actions';
+import {NotificationActionResult, ActionTrigger} from './actions';
 
-import {NotificationOptions, Notification, ActionTrigger, NotificationActionEvent, NotificationClosedEvent, NotificationCreatedEvent} from './index';
+import {NotificationOptions, Notification, NotificationActionEvent, NotificationClosedEvent, NotificationCreatedEvent} from './index';
 
 type Omit<T, K extends keyof any> = Pick<T, Exclude<keyof T, K>>;
